Created leaderboard sprite. Implemented leaderboard for UI and added retrieval of player name to fill out name field on leaderboard. Added fading functionality for displaying and hiding leaderboard using TAB key.

+++ Assets/Scripts/Manager.cs
@@ -9,12 +9,17 @@ public class Manager : MonoBehaviour {
 
 	// Use this for initialization
 	void Start () {
-		PlayerPrefs.DeleteKey ("playerName");
+		Debug.Log (SceneManager.GetActiveScene ().buildIndex);
+		//Check if the scene is the main menu
+		if (SceneManager.GetActiveScene ().buildIndex == 0) {
+			//Delete previously stored player name to create a new player.
+			PlayerPrefs.DeleteKey ("playerName");
+		}
 	}
 	
 	// Update is called once per frame
 	void Update () {
-	
+		
 	}
 
 	//Changes the level to the string named scene defined in the parameter.
@@ -23,11 +28,19 @@ public class Manager : MonoBehaviour {
 	}
 
 	public void startButtonClicked(){
-		playerName = nameInput.text;
-		Debug.Log("Player Name: " + playerName);
-		PlayerPrefs.SetString ("playerName", playerName);
+		//Get and store the new player name.
+		if (nameInput.text == "" || nameInput.text == null) {
+			Debug.Log ("No Input.");
+			//Set a default player name.
+			PlayerPrefs.SetString ("playerName", "Player 1");
+		} else {
+			playerName = nameInput.text;
+			Debug.Log ("Player Name: " + playerName);
+			PlayerPrefs.SetString ("playerName", playerName);
+		}
 		//Change to first level
 		changeLevel (1);
+		
 	}
 
 	public string getPlayerName(){
